refactor comment list
